requirements/production.txt
Django==4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-environ==0.9.0
drf-yasg==1.21.3
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
whitenoise==6.2.0
Pillow==9.2.0
